+++ src/ngdoc.js
@@ -379,7 +379,18 @@ Doc.prototype = {
       }
     });
     flush();
-    this.shortName = this.name.split(/[\.:#]/).pop().trim();
+    var shortName = this.name.split("#");
+    if (shortName.length > 1) {
+      this.shortName = shortName.pop().trim();
+    } else {
+      shortName = this.name.split(":");
+      if (shortName.length > 1) {
+        this.shortName = shortName.pop().trim();
+      } else {
+        this.shortName = this.name.split(".").pop().trim();
+      }
+    }
+    
     this.id = this.id || // if we have an id just use it
       (this.ngdoc === 'error' ? this.name : '') ||
       (((this.file||'').match(/.*(\/|\\)([^(\/|\\)]*)\.ngdoc/)||{})[2]) || // try to extract it from file name
@@ -459,6 +470,8 @@ Doc.prototype = {
           match = text.match(/^([^\s]*)\s+on\s+([\S\s]*)/);
           self.type = match[1];
           self.target = match[2];
+        } else if(atName == 'constructor') {
+          self.constructor = true;
         } else {
           self[atName] = text;
         }
@@ -660,6 +673,9 @@ Doc.prototype = {
 
     dom.h('Usage', function() {
       dom.code(function() {
+        if (self.constructor) {
+          dom.text('new ');
+        }
         dom.text(name.split(':').pop());
         dom.text('(');
         self.parameters(dom, ', ');